+++ frontend/src/validators/definitions/index.ts
@@ -1,0 +1,3 @@
+export { regex } from './regex'
+export { cel } from './cel'
+export { wasm } from './wasm'